:sparkles feat: 커뮤니티 전체글 및 주제별 글 프론트에서 추가해달라했던 데이터까지 추가 완료

+++ hugme-dog/src/main/java/spring/hugme/domain/community/entity/PostHashtag.java
@@ -21,7 +21,7 @@ public class PostHashtag {
 
   @Id
   @GeneratedValue(strategy = GenerationType.IDENTITY)
-  private int hashtagId;
+  private Long hashtagId;
 
   @ManyToOne(fetch = FetchType.LAZY)
   @JoinColumn(name = "post_id")
